+++ mkanak_spring/src/main/java/mkanak_spring/model/services/PostServiceImpl.java
@@ -1,21 +1,39 @@
 package mkanak_spring.model.services;
 
-import mkanak_spring.model.dao.PostDAO;
+
 import mkanak_spring.model.entities.*;
 
+import mkanak_spring.model.filters.PostSpecificationBuilder;
 import mkanak_spring.model.preferences.ViewingPreference;
+import mkanak_spring.model.repositories.*;
 import org.json.simple.JSONArray;
 import org.json.simple.JSONObject;
 import org.json.simple.parser.JSONParser;
 import org.json.simple.parser.ParseException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
+import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
+
 @Service
 public class PostServiceImpl implements PostService{
+
+
     @Autowired
-    PostDAO postDAO;
+    ApartmentRepo apartmentRepo;
+    @Autowired
+    VillaRepo villaRepo;
+    @Autowired
+    PropertyPictureRepo propertyPictureRepo;
+    @Autowired
+    PostRepo postRepo;
+    @Autowired
+    SavedPostsRepo savedPostsRepo;
     @Autowired
     JsonToObject converter;
 
@@ -30,55 +48,55 @@ public class PostServiceImpl implements PostService{
             Villa property = new Villa();
             property = converter.buildVilla(post);
             property.setHasPictures(pictures.size() != 0);
-            postDAO.saveVilla(property);
+            this.saveVilla(property);
             List<PropertyPicture> pictureList = converter.buildPropertyPictures(post, property.getPropertyID());
-            postDAO.saveAllPictures(pictureList);
+            this.saveAllPictures(pictureList);
         }
         else {
             Apartment property = new Apartment();
             property = converter.buildApartment(post);
             property.setHasPictures(pictures.size() != 0);
-            postDAO.saveApartment(property);
+            this.saveApartment(property);
             List<PropertyPicture> pictureList = converter.buildPropertyPictures(post, property.getPropertyID());
-            postDAO.saveAllPictures(pictureList);
+            this.saveAllPictures(pictureList);
         }
     }
 
     @Override
     public List<Post> getHomepagePosts(JSONObject preference,int pageNum,int pageSize) {
         ViewingPreference p = converter.parseViewingPreference(preference);
-        return postDAO.getAllPosts(p,pageNum,pageSize);
+        return this.getAllPosts(p,pageNum,pageSize);
     }
 
     @Override
     public List<Post> getSavedPosts(int id, JSONObject preference,int pageNum,int pageSize) {
         ViewingPreference v = converter.parseViewingPreference(preference);
-        return postDAO.getSavedPostsByUserID(id,v,pageNum,pageSize);
+        return this.getSavedPostsByUserID(id,v,pageNum,pageSize);
     }
 
     @Override
     public List<Post> getProfilePosts(int targetUserID, JSONObject preferences,int pageNum,int pageSize) {
         ViewingPreference v = converter.parseViewingPreference(preferences);
-        return postDAO.getPostsByUser(targetUserID,v,pageNum,pageSize);
+        return this.getPostsByUser(targetUserID,v,pageNum,pageSize);
     }
 
     @Override
     public List<Long> getSavedPostsIDs(int userID) {
-        return postDAO.getSavedPostsIDs(userID);
+        return this.getSavedPostsIDs(userID);
     }
 
     @Override
     public void addToSavedPosts(JSONObject saveEntry) {
         int userID = (int) saveEntry.get("userID");
         int postID = (int) saveEntry.get("postID");
-        postDAO.addToSavedPosts(userID,postID);
+        this.addToSavedPosts(userID,postID);
     }
 
     @Override
     public void removeFromSaved(JSONObject entry) {
         int userID = (int) entry.get("userID");
         int postID = (int) entry.get("postID");
-        postDAO.removeFromSavedPosts(userID,postID);
+        this.removeFromSavedPosts(userID,postID);
     }
 
     @Override
@@ -93,7 +111,7 @@ public class PostServiceImpl implements PostService{
 
     @Override
     public JSONObject getPostDetails(long postID) {
-        return postDAO.getPostDetails(postID);
+        return postRepo.getPostLargeView(postID);
     }
 
 
@@ -117,6 +135,85 @@ public class PostServiceImpl implements PostService{
 
 
 
+
+
+
+    public void saveApartment(Apartment property) {
+        apartmentRepo.save(property);
+    }
+
+    public void saveVilla(Villa property) {
+        villaRepo.save(property);
+    }
+
+    public void saveAllPictures(List<PropertyPicture> propertyPictureList) {
+        propertyPictureRepo.saveAll(propertyPictureList);
+    }
+
+
+
+    public List<Post> getAllPosts(ViewingPreference preference, int pageNum, int pageSize) {
+        PostSpecificationBuilder pb;
+        if(preference!=null && preference.isFiltered()
+                && Objects.equals(preference.getFilterPreference().getPropertyType(), "apartment")
+                && preference.getFilterPreference().isStudentHousing()) {
+
+            List<Long> studentHouseIDs= apartmentRepo.getStudentHousingIDs(preference.getFilterPreference().isStudentHousing());
+            pb = new PostSpecificationBuilder(preference,studentHouseIDs);
+        } else {
+            pb = new PostSpecificationBuilder(preference);
+        }
+        return buildAndReturn(pb,pageNum,pageSize);
+    }
+
+
+    private List<Post> buildAndReturn(PostSpecificationBuilder pb,int pageNum,int pageSize){
+        Specification<Post> sps = pb.build();
+        Sort s = pb.getSort();
+        if(sps==null && s == null) {
+            Pageable page = PageRequest.of(pageNum, pageSize);
+            return postRepo.findAll(page).toList();
+        }
+        else if(s==null) {
+            Pageable page = PageRequest.of(pageNum, pageSize);
+            return postRepo.findAll(sps,page).toList();
+        }
+        else if(sps==null) {
+            Pageable page = PageRequest.of(pageNum, pageSize,s);
+            return postRepo.findAll(page).toList();
+        }
+        else {
+            Pageable page = PageRequest.of(pageNum, pageSize,s);
+            return postRepo.findAll(sps, page).toList();
+        }
+    }
+
+    public List<Post> getPostsByUser(int id,ViewingPreference preference, int pageNum,int pageSize){
+        PostSpecificationBuilder pb = new PostSpecificationBuilder(preference,id);
+        return buildAndReturn(pb,pageNum,pageSize);
+    }
+
+
+    public List<Post> getSavedPostsByUserID(int id,ViewingPreference preference, int pageNum,int pageSize){
+        List<Long> postIDsSaved = savedPostsRepo.getUserSavedPostsIDs((long) id);
+        PostSpecificationBuilder pb = new PostSpecificationBuilder(preference,postIDsSaved);
+        return buildAndReturn(pb,pageNum,pageSize);
+    }
+
+    public void removeFromSavedPosts(long userID, long postID){
+        savedPostsRepo.deleteSavedPost(userID,postID);
+    }
+
+    public void addToSavedPosts(long userID, long postID){
+        SavedPostsEntry entry = new SavedPostsEntry();
+        entry.setPostID(postID);
+        entry.setUserID(userID);
+        savedPostsRepo.save(entry);
+    }
+
+    public List<Long> getSavedPostsIDs(long id){
+        return savedPostsRepo.getUserSavedPostsIDs(id);
+    }
 
 
 }
