+++ mkanak_spring/src/main/java/mkanak_spring/model/services/UserServiceImpl.java
@@ -2,28 +2,29 @@ package mkanak_spring.model.services;
 
 import mkanak_spring.model.entities.User;
 import mkanak_spring.model.entities.UserCredentials;
-import mkanak_spring.model.dao.UserDAO;
 import mkanak_spring.model.repositories.UserCredentialsRepo;
+import mkanak_spring.model.repositories.UserRepo;
 import org.json.simple.JSONObject;
 import org.json.simple.parser.ParseException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import java.util.List;
 
 @Service
 public class UserServiceImpl implements UserService {
     @Autowired
-    UserDAO userDAO;
-    @Autowired
     UserCredentialsRepo userCredentialsRepo;
+
+    @Autowired
+    UserRepo userRepo;
+
     @Override
     //@Transactional
     public Long createUser(JSONObject user) throws ParseException {
         JsonToObject converter = new JsonToObject();
         User userInstance = converter.getUserFromJson(user);
         userInstance.setUserID(null);
-        if(userDAO.usernameExists(userInstance.getUsername())) return -1L;
-        userDAO.saveUser(userInstance);
+        if(this.usernameExists(userInstance.getUsername())) return -1L;
+        this.saveUser(userInstance);
         return userCredentialsRepo.findByUsername(userInstance.getUsername()).getUserID();
     }
 
@@ -32,24 +33,38 @@ public class UserServiceImpl implements UserService {
     public Long logInUser(String username, String password) {
         JsonToObject loginManager = new JsonToObject();
         UserCredentials userCredentials = loginManager.SignInUser(username, password);
-        userCredentials = userDAO.logInUser(userCredentials);
+        userCredentials = this.logInUser(userCredentials);
         if(userCredentials == null) return -1L;
         if(!userCredentials.getPassword().equals(password)) return -2L;
         return userCredentials.getUserID();
     }
     @Override
     public User findUserInfoByUseId(long userID) {
-        return userDAO.findUserInfoByUseId(userID);
+        return userRepo.findUserInfoByUseId(userID);
     }
-//    @Override
-//    public String findUserPhoneByUseId(long userID) {
-//        return userDAO.findUserPhoneByUseId(userID);
-//    }
+
+
+    public void saveUser(User user) {
+        userRepo.save(user);
+    }
+    public UserCredentials logInUser(UserCredentials userCredentials) {
+        userCredentials = userCredentialsRepo.findByUsername(userCredentials.getUsername());
+        return userCredentials;
+    }
+
+    public Boolean usernameExists(String username) {
+        return userRepo.findByUsername(username);
+    }
+
+    public Boolean emailExists(String email) {
+        return userRepo.findByEmail(email);
+    }
+
+
 
     @Override
     public boolean editUser(JSONObject userJson)  {
         // TODO validate user
-
         JsonToObject converter = new JsonToObject();
         User user = null;
         try {
@@ -57,7 +72,12 @@ public class UserServiceImpl implements UserService {
         } catch (ParseException e) {
             return false;
         }
-        userDAO.saveUser(user);
+        this.saveUser(user);
         return true;
     }
 }
+//    @Override
+//    public String findUserPhoneByUseId(long userID) {
+//        return userDAO.findUserPhoneByUseId(userID);
+//    }
+
